refactor: 💡 move Dirent into a separate file
