+++ src/__tests__/volume.test.ts
@@ -1,6 +1,7 @@
 import { URL } from 'url';
-import { Link, Node, Dirent } from '../node';
+import { Link, Node } from '../node';
 import Stats from '../Stats';
+import Dirent from '../Dirent';
 import { Volume, filenameToSteps, StatWatcher } from '../volume';
 
 describe('volume', () => {
