+++ src/index.ts
@@ -1,5 +1,5 @@
-import { Dirent } from './node';
 import Stats from './Stats';
+import Dirent from './Dirent';
 import { Volume as _Volume, StatWatcher, FSWatcher, toUnixTimestamp, IReadStream, IWriteStream } from './volume';
 import * as volume from './volume';
 import { IPromisesAPI } from './promises';
