+++ src/volume.ts
@@ -1,7 +1,8 @@
 import { resolve as resolveCrossPlatform } from 'path';
 import * as pathModule from 'path';
-import { Node, Link, File, Dirent } from './node';
+import { Node, Link, File } from './node';
 import Stats from './Stats';
+import Dirent from './Dirent';
 import { Buffer } from 'buffer';
 import setImmediate from './setImmediate';
 import process from './process';
